+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Server.scala
@@ -1,9 +1,11 @@
 package io.buoyant.linkerd
 
+import com.twitter.conversions.time._
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.concurrent.AsyncSemaphore
 import com.twitter.finagle.Stack.{Param, Params}
 import com.twitter.finagle.filter.RequestSemaphoreFilter
+import com.twitter.finagle.service.TimeoutFilter
 import com.twitter.finagle.ssl.Ssl
 import com.twitter.finagle.transport.Transport
 import com.twitter.finagle.{ListeningServer, Stack}
@@ -80,20 +82,21 @@ class ServerConfig { config =>
   var tls: Option[TlsServerConfig] = None
   var label: Option[String] = None
   var maxConcurrentRequests: Option[Int] = None
+  var timeoutMs: Option[Int] = None
 
   private[this] def requestSemaphore = maxConcurrentRequests.map(new AsyncSemaphore(_, 0))
 
   @JsonIgnore
   protected def serverParams: Stack.Params = Stack.Params.empty
-    .maybeWith(tls.map {
-      case TlsServerConfig(certPath, keyPath) => tlsParam(certPath, keyPath)
-    }) + RequestSemaphoreFilter.Param(requestSemaphore)
+    .maybeWith(tls.map(tlsParam(_)))
+    .maybeWith(timeoutMs.map(t => Router.ServerTimeout(t.millis))) +
+    RequestSemaphoreFilter.Param(requestSemaphore)
 
   @JsonIgnore
-  private[this] def tlsParam(certificatePath: String, keyPath: String) =
-    Transport.TLSServerEngine(
-      Some(() => Ssl.server(certificatePath, keyPath, null, null, null))
-    )
+  private[this] def tlsParam(config: TlsServerConfig) =
+    Transport.TLSServerEngine(Some(
+      () => Ssl.server(config.certPath, config.keyPath, null, null, null)
+    ))
 
   @JsonIgnore
   def mk(pi: ProtocolInitializer, routerLabel: String) = Server.Impl(